RUBY-1395 Fix auto-restarting thread in after_fork
The changes to properly handle thread restarts for Resque inadvertantly
broke the restarting in more standard cases for transaction finishes
starting up a threads like we rely on for Unicorn.

To address this, we now have NewRelic::Agent::Agent#after_fork perform
locking around the harvester's mark_started check instead of Harvester
and dispense with the worker thread nil and alive? checks that were
making Harvester try to clear out the worker thread on a restart.

Conflicts:
	lib/new_relic/agent/agent.rb