RUBY-1410 Make AgentHooks middleware safer in the face of weird apps
Applications may hand a reference to the Rack env hash off to a background
thread, and iterate over it from that background thread, which was causing
issues when the AgentHooks middleware was trying to write into the env hash
after calling through to the next middleware.

Instead of writing in the env hash twice, we now only write once, before calling
through to the middleware layer below us, which makes it much less likely that
anyone has had a chance to pass the Rack env hash off to a background thread yet
(though it's not impossible - a middleware above us could do this in theory, but
at least it would show up in the backtrace if it did).